requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pyparsing==2.4.7
PyDrive==1.3.1
google-api-python-client==2.103.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
numpy==1.24.3
scipy==1.10.1
opencv-python-headless==4.8.1.78
torch==2.2.0
torchvision==0.17.0
efficientnet-pytorch==0.7.1
Pillow==10.0.0
librosa==0.10.1
matplotlib==3.7.2
soundfile==0.12.1
gunicorn==21.2.0
python-dotenv==1.0.1
--extra-index-url https://download.pytorch.org/whl/cpu
